Consolidate common tests, template specialize CLI
