+++ src/tests/CMakeLists.txt
@@ -13,30 +13,12 @@ endif (COMPILE_UNIT_TESTS)
 
 if(NOT MSVC)
   # Testing apps
-  set(APP_TESTS)
-  add_executable("bergamot-test-native" "native-cli.cpp")
-  
-  if(CUDA_FOUND)
-    target_link_libraries("bergamot-test-native" bergamot-translator)
-  else(CUDA_FOUND)
-    target_link_libraries("bergamot-test-native" bergamot-translator)
-  endif(CUDA_FOUND)
-  
-  set_target_properties("bergamot-test-native" PROPERTIES RUNTIME_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}")
+  set(TEST_BINARIES native wasm intgemm-resolve)
+  foreach(binary ${TEST_BINARIES})
+      add_executable("${binary}" "${binary}.cpp")
+      target_link_libraries("${binary}" bergamot-translator)
+      set_target_properties("${binary}" PROPERTIES RUNTIME_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}/tests/")
+  endforeach(binary)
 
-  add_executable("bergamot-test-wasm" "wasm-cli.cpp")
-  
-  if(CUDA_FOUND)
-    target_link_libraries("bergamot-test-wasm" bergamot-translator)
-  else(CUDA_FOUND)
-    target_link_libraries("bergamot-test-wasm" bergamot-translator)
-  endif(CUDA_FOUND)
-  
-  set_target_properties("bergamot-test-wasm" PROPERTIES RUNTIME_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}")
-
-  # Adding an intgemm_resolve cmdline
-  add_executable(intgemm-resolve intgemm_resolve.cpp)
-  target_link_libraries(intgemm-resolve PRIVATE bergamot-translator)
-  set_target_properties(intgemm-resolve PROPERTIES RUNTIME_OUTPUT_DIRECTORY "${CMAKE_BINARY_DIR}")
 endif(NOT MSVC)
