+++ src/tests/native.cpp
@@ -1,4 +1,4 @@
-#include "apps.h"
+#include "common.h"
 #include "translator/parser.h"
 #include "translator/service.h"
 #include "translator/translation_model.h"
